Even enemies should spawn relatively evenly feat

+++ ProjectFlightSchool/ProjectFlightSchool/Player.h
@@ -44,8 +44,8 @@ class Player: public RemotePlayer
 		bool		mHasMeleeStarted;
 		bool		mLock;
 		bool		mCloseToPlayer;
-		int			mXP;
-		int			mNextLevelXP;
+		float		mXP;
+		float		mNextLevelXP;
 		int			mCurrentLevel;
 		int			mCurrentUpgrades;
 
@@ -114,7 +114,7 @@ class Player: public RemotePlayer
 
 	protected:
 	public:
-		void		AddXP( int XP );
+		void		AddXP( float XP );
 		void		PickUpEnergyCell( EnergyCell** energyCell );
 		void		DropEnergyCell( EnergyCell** energyCells );
 		void		GiveEnergyCellToShip( EnergyCell** energyCells, UINT shipID, DirectX::XMFLOAT3 shipPos );
